requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
Django==1.10.6
django-celery-beat==1.0.1
djangorestframework==3.6.2
kombu==4.0.2
packaging==16.8
pyparsing==2.2.0
pytz==2016.10
redis==2.10.5
six==1.10.0
vine==1.1.3
